MINOR: quic_tls: Add more definitions for QUIC-TLS.
Note that as QUIC-TLS do not use at all the record layer of TLS,
we have to implement the higher layer part of TLS to be reused for
our QUIC-TLS implementation, unless there exists a TLS API whis allows us
to get rid of its record layer part. Here are the first useful definitions.