MINOR: quic: Add new data structure used for QUIC-TLS.
Add some definitions of structure for the QUIC-TLS packet number space
("quic_pkt_ns" struct) and attach such a data struct for each QUIC-TLS
context and for each direction of transmission (tx/rx).
Add QUIC_TLS_TAG_LEN(16)macro: according to the RFC for QUIC-TLS, there is
a 16 bytes fixed size for the authentication tags of the QUIC-TLS ciphersuites.