MINOR: quic: Modify the API to decode QUIC variable-length integers.
This patch only modifies the prototype of quic_dec_int() function so that
to pass the address of the value to store the decoding result as parameter
in place of returning it. This is to ease the code to come especially the
QUIC frame parsing part which has to decode a lot of variable-length integers.