MINOR: quic: Replace QUIC frame type macros by an enum.
We replace QUIC_FT_* frame types macros by a C enum.
We also add QUIC_FT_MAX new value for this enumeration to ensure we do not call
a undefinde QUIC frame parser.
Furthermore this patchs adds a function to display the name all the parsed QUIC
frame (debug traces).
In the PADDING frame parser we modified the condition to stop parsing such a frame.
We do not stop the parsing when the current pointer in the buffer reaches the end
of this buffer (*buf == end) but while the pointer is below the end (*buf < end).
This is more secure if we want to call this parser several times with the same buffer
even if this will never be the case for QUIC: if present, a PADDING frame is always
the last frame of a QUIC packet.