BUG/MAJOR: quic: Wrong QUIC connection ID storage.
Replace the ebtree to store QUIC connection IDs by an ebmbtree (in place of an
ebpttree which did not make sense).
Furthermore this path takes into an account that an Initial packet may be sent
several times by a client to a server. Such an incoming packet identifies the
connection thanks to the random destination connection ID choosen by the client.
This destination connection ID is replaced by the server. This is the definitive
connection ID used to match the client connections just after having received the
Initial packet.