MINOR: quic: Make the QUIC packet reader handle all the packet.
This simple patch renames quic_packet_read_header() to quic_packet_read() to
reflect its job: read a packet (not only the header).
It also modifies the value of the offset in the UDP datagram used to point to the
next QUIC packet which is passed to this function so that to read all the QUIC
packets. It finally updates quic_packets_read() so that to call quic_packet_read()
until having consumed all the UDP datagram payload.