MINOR: ssl: Rework of ssl_sock_init().
This patch extracts the code responsible of initializing the BIO and SSL sessions
objects of all SSL connections so that to reuse it elsewhere later, for instance
for QUIC connections which only needs SSL and BIO objects at the TLS layer stack
to work. This is the job of the new ssl_bio_and_sess_init() function implemeted
by this patch.

Rework a little ssl_sock_init() to make it more maintainable. After having called
ssl_bio_and_sess_init() the remaining things to do are:
- for a client part:
   - initialize an SSL_SESSION object,
   - set the connect state of the SSL session.
- fot the listener part:
   - allocate the 0RTT data buffer,
   - set the accept state of the SSL session.

Then both have to set their flags and start the handshake.