MINOR: quic-tls: QUIC TLS API modifications for AEAD protection.
Move quic_payload_encrypt() and quic_payload_decrypt() from src/quic_conn.c
to src/quic_tls.c and rename to quic_tls_encrypt() and quic_tls_decrypt.
Furthermore modifies the API to be able to pass the Additional Authentication Data
as parameter to quic_tls_encrypt() as this is done for the decryption part.