MINOR: quic-tls: Improve the QUIC TLS API reasdibilty.
To make them reusable afterwards, this patch modifies the QUIC TLS function
prototypes of quic_tls_encrypt() and quic_tls_decrypt() renaming the
variables passed as parameters to reflect the needs of the underlying cryptographic
functions used for AEAD to encrypt and decrypt the QUIC packets. What was not
clear is the AAD parameter required for AEAD. This is why a ->aad_len is added
to the quic_packet struct to deduce the AAD to pass to AEAD when emplementing
quic_tls_encrypt. This aad_len parameter is stored just after having remove the
QUIC packts header protection. Furthermore after having removed this protection
the ->len value of a quic_packet struct is no more the length field value of the
QUIC packet which includes only the remaining data length after the packet number
field, but the real length of the QUIC packet, header included, so AAD included.
With such information, the AAD length and the QUIC packet full length, we will be
able to ask to a tasklet to decrypt the QUIC packet data of quic_packet structs.