MINOR: quic: Enable the usage of draft27 transport parameters encoding.
This patch modifies the initialization of the quic_conn struct upon incoming
new connection to really store the QUIC version at the rigth place. This latter is
reused when initializing the QUIC connection I/O handler tasklet. This works with
picoquic which support draft27 contrary to ngtcp2 at the time of this commit.