MINOR: quic: Store the information of TX CRYPTO frames.
We store the CRYPTO frames which are sent for each encryption level ordered by
packet number. Note that we build at most one CRYPTO frames by handshaked packet.
As the packet are acknowledged or declared as lost thanks to their packet number,
It will be easy to retransmit such CRYPTO frames if they are detected to be lost or to remove
them from the remaining CRYPTO frames to be acknowledged if they have been acknowledged.