BUG/MINOR: quic: Missing output buffer reset.
Each time we enter the function which send/retransmit CRYPTO frames, if there is
no new frame to prepare we must check if there is remaining data to send in the
output buffer. If this is the case, after having sent these data we must no forget
to reset the output buffer if we do not want to send again the same data.