MINOR: quic: Add a new I/O handler for client connections.
This is a very simplistic patch to implement another I/O handler for the client
connections. At this time it behaves the same way as the I/O handler for listeners.
The difference is that the underlying fd owner is the connection opened for the client
in place of the listener.