MINOR: quic: Enhance client Handshake packets decryption.
Before trying to decrypt any handshake packet, we remove the header protection
of those which could not be unprotected because the secrets were not available.
When done, as their packet numbers are readable, we put them in the tree of packets
ordered by packet number so that to be decrypted.