BUG/MINOR: quic: Wrong encryption level switching handling.
When entering Handshake level, it is possible that the last Initial
packet which was sent has not fulfill a buffer (or a UDP datagram).
With this patch we consume a packet buffer only if there is no
more Handshake level data to send. This is the client for a client
which only sent Initial packets and wait for the server response.
On the server side, there is always Handshake data to send after
leaving the Initial level.