MINOR: quic: Prepare the retransmission rework.
This is a very simple patch to prepare the retransmission feature which has been
intentionaly broken by 0cc45dd commit to make the code be more readable.
We add a parameter to quic_build_handshake_packet() and quic_do_build_handshake_packet()
to pass the length of the CRYPTO data to build a handshake packet. For retransmission,
this will be the length of TX CRYPTO frames which should be passed (see quic_tx_crypto_frm
struct definition).