MINOR: quic: Inconditionaly treat the writter buffer.
So that not to reach corner cases when preparing/writing outoing QUIC packets,
make q_next_wbuf() returns the next buffer to be written. This is the responsability
of the caller to know if the returned buffer may be used to build the next packet.
Also add a boolean function to know if a QUIC buffer is empty or not.