BUG/MINOR: quic: Last handshake packet not sent.
There are cases depending on the maximum of crypto data in flight size and the number
to handshakes to send which lead to bugs where the last handshake packets may not be sent.
This is due to the API of the QUIC buffers (q_buf struct) which was not safe.
Make usage of the safer version of q_next_wbuf() from previous commit to
detect the writability of a writter buffer (a buffer to be written to build
outgoing packets). For now on a buffer is writable if not empty or flagged
as to be writable. As the reader buffer reset (or empty) the buffer it has successfully
read, the writability may be safely detected this way.