MINOR: quic: QUIC specific traces.
To be able to use the trace API (proto/trace.h) from other source files
than the one TRACE_SOURCE macro is defined, we must use our own wrapper
around this API. The letter Q (for QUIC) is prepended to any TRACE_*() functions.
This is also the case for TRACE_SOURCE macro. Furthermore all the QUIC_EV_CONN_*
macro for the trace events must be exported.