MINOR: quic: Add some robustness to max_stream_data_size().
This function computes the maximum number of bytes we can put in a buffer.
Do not rely on the caller to make correct checks on passed parameters to
this function and modify this function to check the passed parameter
values to avoid buffer under/overflows. Rename <len> passed parameter
to <ilen> to avoid some confusion with the internal <len_sz> variable.
Modify max_available_room() to make it more understandable.