MAJOR: quic: QUIC handshake packet building rework.
When building handshake packets, the previous code naively built handshake
packets with at most one CRYPTO frame by packets. This is not efficient during
retransmission where we are supposed to be able to send CRYPTO frames which are
not continguous. Furthermore, when the TLS stack provide us with CRYPTO data to
send, we build CRYPTO frames which are sequentially put in a list to be send.
Then we build handshake packets from this list and keep traces of the CRYPTO
frames which have been sent. When retransmission are needed, we put back the
CRYPTO frames which have been lost in the list of CRYPTO frame to send, and so on.
So this is the same function which builds the packet to send or to retransmit.
The function for retransmission has been removed. Same thing for its associated
trace event definitions.