MINOR: quic: Support for CRYTO frames of any length.
Modify the CRYPTO frame builder so that it does not suppose any more
that the CRYPTO lengths are not bigger than a prebuilt buffer allocated
when the TLS stack provides us with CRYPTO data. So we will have less
chances to build packets with several CRYPTO frames, except obviously
when they are not contiguous.