MINOR: quic: Allocate as little as possible TX CRYPTO frames.
When the TLS stack provides us with CRYPTO data, it provides use with the
length of this CRYPTO data. As our CRYPTO frame builder is able to build a
frame of any length, no matter if the length provided by the TLS stack is larger
than our underlying buffer to store these data, we allocate a TX CRYPTO frame
with the length provided by the TLS stack as length.