MINOR: quic: Do not distinguish Application RX packets for handshake packets.
The aim of this patch is to process the Application level QUIC packet the same
way as the handshake packets remove quic_treat_packets() function which
was there to process only Application level packets.
Remove quic_tx_crypto_frm struct and make usage of quic_tx_frm struct
which is made of a union for all the frames for both CRYPTO and Application level
only frames (Support only NEW_CONNECTION_ID and CRYPTO frames at this time)
as this is at this time the frames supposed to be retransmitted when lost.