BUG/MAJOR: quic: TX Packet numbers must be stored as int64_t.
To be decoded, truncated packets numbers are compared to the largest acked packet number
which is stored as int64_t and initialized with -1 value. When compared with
a uint64_t TX packet number this leaded this largest acked packet number  to never
be incremented. Consequently with a 1 byte length, decoded packet numbers were
wrapped after every 256 packets.