MINOR: quic: Update the loss detection information upon ACK frame receipt.
We must update the loss information when receiving ACK frames if the largest
acknowledged packet is newly acknowledged and if there is at least one ack-eliciting
packet which is newly acknowlegded. So we accumulate the acked TX packets flags
into qc_ackrng_pkts() function new <pkt_flags> passed paratemer. We must make this
function able to start looking for the largest acked packet as optional parameter
(see new <largest_node> parameter) which is initialized only if this packet
was newly acknowledged.