MINOR: quic: Add time_of_last_eliciting to loss detection information.
Add ->time_of_last_eliciting to quic_pktns struct to store the time the ack-elicinting
packet has been sent.
Als add <len> member to quic_tx_packet struct to store the length of TX packets
which are in-flight.
Make a single call to usec_now() for all the packets belonging to the same sent
UDP datagram.