BUG/MINOR: quic: Wrong ACK frame parsing for reordered packets.
When a range of acked packets must be treated, if a packet has been reordered,
it is possible that less recent packets has already been
treated and removed from the tree of TX packets. In this case the largest
node of the range cannot be found and we must start from the  first node among
this range which is still present in the tree in decreasing order.