BUG/MAJOR: quic: Non in flight packet could not be sent anymore.
Non in-flight packets (packets which do not contain ack-eliciting frames or PADDING
frames) coult not be sent anymore. This is the case for ack-only packets which
are oftend sent by clients. This was due to this commit:
  "MINOR: quic: Add time_of_last_eliciting to loss detection information."
which prevent the ->data field of prepared buffer to be initialized.
This patch also renames ->len of quic_tx_packet struct to ->in_flight_len which
is null for non in-flight packets.