BUG/MINOR: quic: Fix the case where the firts Initial packet from server is lost.
In this case, the client resend an Initial packet with the same random as DCID.
The listener must look for the connection in ->icids tree. If this listen does not
have anymore packet to send for any reason (anti-amplification limit reached or
because of the congestion control), the clien may reply with an Initial packet
but with the DCID the server has choosen to send its first packets. In this case
we must look for the CID in ->cids tree.
This must be done when the packet type is Initial, <cids> value is ->icids, the
length of the DCID equals the one choosen by the server, and if the node could
not be found. To know how to retrieve the quic_conn struct from this node if found
we set <cids> to NULL to distinguish this special case from the first one where
the DCID node were choosen by the client. So, in this case we must use ->scid_node
node entry of the quic_conn struct as for the non Initial packets.