MINOR: quic: Add counters from UDP datagram payload counting.
We must count the bytes sent/received by UDP datagram, especially
for servers which must not send more than three times bytes as received
before the peer address has been validated.
Make usage of a datagram context struct for that, used for each packet
found in a datagram. Put the DCID node in this struct to be matched
between two packets in a UDP datagram with the first one.
Make usage of the new tx.bytes and rx.bytes counters of quic_conn struct
each time a connection could be identified for a UDP datagram regardless
of the packet parsing statuses. There is nothing complicated to do on the
sending part, only incrementing the tx.bytes counter by the size of the
sent datagram size.