BUILD: ssl_sock: Remove duplicate ssl_sock_switchctx_cbk() definitions.
There is duplicated definitions ssl_sock_switchctx_cbk() which have been
committed by this patch:
    "MINOR: ssl: Export definitions required by QUIC".
I guess the current patch should be merged in the mentionned one above.
Furthermore there is a case where this function is static (when BoringSSL
is not used and when HA_OPENSSL_VERSION_NUMBER < 0x10101000L).