BUG/MAJOR: quic: Possible infinite loop when building handshake packets.
It is possible that qc_do_build_hdshk_pkt() returns -1 indicating
the remaining room in the TX passed buffer is not big enough to
build a handshake packet because of the congestion control window
limit. In such a case we must not consume a buffer and select the next
one if the consumed buffer was not supposed to be reused. This would
make this sequence repeat endlessly.